requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==1.10.8
mysql-connector-python==8.0.33
pyodbc==4.0.39
starlette==0.30.0
